Update Api to use same signature as WinRT
`void NavigateToLocalStreamUri(String, IUriToStreamResolver)` -> `void NavigateToLocalStreamUri(Uri, IUriToStreamResolver)`.

Update to take `Uri` due to protections that could be made for defensive programing. The parameter is validated to ensure it is not null and a relative URI. After basic validation the `UriHelper` class is used to pack and unpack the URI to ensure we are not exposed to additional attack vectors.