Protect actions and delay navigation until last moment
An issue may occur where methods, such as InvokeScript/InvokeScriptAsync, are permitted to be invoked prior to the WebView completing initialization. As such, properties may not yet be set for permitting JavaScript or ScriptNotify, causing inconsistencies with application behavior.

Additionally, the Source property is set, causing an immediate navigation before other properties are also set. The code has been updated for WPF to follow a similar order for WinForms.

See #2346, #2347