Add Close before HWND destroyed
During the disposal process the HWND for the control is destroyed, leaving an orphaned HWND core window. In RS4 this could case a lock up since it was expected that the Close method on WebViewControl would always be closed. During RS5 this was updated to ensure that Close is always called, even if the hosting process crashes or one of the supporting processes is terminated unexpectedly.

This change intercepts the DestoryHandle and OnHandleDestoryed methods of the Windows Forms WebView control to first call Close method. This starts a synchronous shutdown of all processes and unparents the HWNDs. By the time the actual disposal comes around, the control and its processes are already cleaned up and the UI remains responsive.

Note that this change is to work around a RS4 bug; however, its inclusion for RS5 does not impede compatibility or introduce any undesired consequences.