Add support for creating web views in the same WWAHost process
With the traditional API, invoking `new WebView()` would always create a web view in a new WWAHost and Win32WebViewHost process. In some cases, state needs to be shared between individual views, such as for single sign on scenarios, to avoid having each web view authenticate.  Edge does not persist session state to disk to both maintain performance and provide a safe way to remove session state in the event of a crash. When a new web view process is created, Edge internally copies all the session state it can to the new process. With the default .ctor each web view was created in its own WWAHost process, and each was treated as a new session, failing to inherit any of the ephemeral session state. However, any state that was persisted to disk, such as cookies with a future expiration, would be read into the new web view session.

The new API permits passing in a `WebViewControlProcess` via .ctor overload that will be used to create the process. An instance of `WebViewControlProcess` can be obtained by calling `new WebViewControlProcess()` or by using the `WebView.Process` property of an existing web view. Tests have also been added to ensure cookie behavior is maintained between web views in the same process, and thus the same session, but not web views that are in different processes (and thus in different sessions).

Resolves #2375