Swap NuGet package references for System.Runtime.WindowsRuntime
Packages System.Runtime.WindowsRuntime and System.Runtime.WindowsRuntime.UI.Xaml swapped with assembly versions shipped with .NET framework. Note that the assemblies do not show on the reference assembly browser and cannot be resolved by name alone. A hint path has been provided with the location of the assemblies on disk and is necessary for them to resolve.