XamlIslands: Winforms activation / focus fix
When moving focus, use the XamlSourceFocusNavigationRequest for correlation identifier. In some cases there may be an issue where focus is required when there is only one UWP element in an island. As a workaround for a UWP platform issue, the correlation identifier can be used to determine if a chain of focus is occurring as a result of WndProc or ProcessTabKey input.