Xaml Islands: Handle native DLL load attempt
When attempting to load an EXE or DLL that cannot be loaded by the CLR (most common case is the item is NATIVE, not CLR, or that the item is for a different bitness), a `BadImageFormatException` may occur. Since the `MetadataProviderDiscovery` class is best-effort, catch the exception and continue processing the types and assemblies.