Ensure WPF WebView initialized before HWND pos
During initialization the size of the control may not be correctly set causing a grey rect to be displayed. This is due to a timing issue with the dispatcher wherein the control has been resized by the visual layout system but has not yet finished initialization, causing the bounds to not be set.

The call to update the render size has been moved to after the control is initialized to resolve this issue. It causes a noticeable layout event wherein the grey box is displayed then swapped for the actual control.