Reset horizontal pos during WPF WebView bounds update
The vertical and horizontal coordinates are relative to the Window, not the containing UIElement. Given a Window with a Grid of only three columns, each 200px wide, the WebView set to Grid position 0,1 X value of 200, placing it in the third column of the Grid instead of the expected second column.

There was already a hack in place to zero out the Y position of the location. This has been extended to the X position as well.