Fall back to WebBrowser controls when running as admin
The Edge WebView does not at the time of this writing (28. march 2020)
support running in an application with elevated permissions. This commit
adds a test to `WebViewControlHost.IsSupported` which checks if the
application is running as an administrator so that `WebViewCompatible`
will fall back to using the legacy WebBrowser controls.

The tests will unfortunately only work when running the test suite as
administrator, and that will make all the other tests fail.  The test
will therefore be marked as inconclusive instead of failling when the
test suite is running as a non elevated user so that we won't always
have at least one breaking test when we run the test suite.